requirements.txt
requests
scikit-learn
tqdm
progressbar
albumentations
imgaug
imageio
matplotlib
pillow  
opencv-python
numpy 
python-magic
pandas